+++ test-compute-kernels/CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [1.4.0] - 2022-12-15
+### Added
+### Changed
 - Update conanfile requirements to use version-ranges
 ### Deprecated
 ### Removed